Resultado decimal

+++ Ejercicio6.java
@@ -4,11 +4,11 @@ public class Ejercicio6 {
     public static void main(String[] args) {
         Scanner entrada = new Scanner(System.in);
         System.out.print("DIgite un numero: ");
-        int n1= entrada.nextInt();
+        double n1= entrada.nextInt();
         System.out.print("DIgite otro numero: ");
-        int n2= entrada.nextInt();
+        double n2= entrada.nextInt();
         System.out.print("DIgite otro numero: ");
-        int n3= entrada.nextInt();
+        double n3= entrada.nextInt();
         entrada.close();
         double media=(n1+n2+n3)/3;
         System.out.println("Media aritmetica = "+media);
